requirements.txt
absl-py==1.1.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
comm==0.1.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
dotmap==1.3.30
einops==0.4.1
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.33.3
freetype-py==2.3.0
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
importlib-resources==5.12.0
inplace-abn==1.1.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter-client==8.1.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-packaging==0.12.3
jupyter-server==2.5.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.3
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.21.0
jupyterlab-widgets==3.0.6
kiwisolver==1.4.3
lpips==0.1.4
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.3.4
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.2
multiprocess==0.70.12.2
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.4
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.2.2
open3d==0.13.0
opencv-python==4.5.3.56
opencv-python-headless==4.6.0.66
openmesh==1.2.1
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
plyfile==0.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pyglet==2.0.5
Pygments==2.14.0
PyOpenGL==3.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrender==0.1.45
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.7.2
pytz==2022.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.12.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tenacity==8.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.7
torch==1.11.0+cu113
torchmetrics==0.9.1
torchvision==0.12.0+cu113
tqdm==4.62.3
traitlets==5.9.0
trimesh==3.9.41
typing-extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.2